chore: Rename story data file
