+++ client/web/src/enterprise/batches/preview/BatchChangePreviewPage.story.tsx
@@ -19,7 +19,7 @@ import {
 
 import { BATCH_SPEC_BY_ID } from './backend'
 import { BatchChangePreviewPage } from './BatchChangePreviewPage'
-import { visibleChangesetApplyPreviewNodeStories, hiddenChangesetApplyPreviewStories } from './list/testData'
+import { visibleChangesetApplyPreviewNodeStories, hiddenChangesetApplyPreviewStories } from './list/storyData'
 
 const decorator: DecoratorFn = story => <div className="p-3 container">{story()}</div>
